requirements.txt
torch
torchvision
scikit-image
pillow
numpy
matplotlib
scipy
